organize helpers into utils folder (#751)
