+++ tests/e2e/pages/home.spec.ts
@@ -1,5 +1,5 @@
+import { e2eA11y, logInAdminUser, logInStateUser } from "../utils";
 import { test, expect } from "../utils/fixtures/base";
-import { logInStateUser, logInAdminUser, e2eA11y } from "../helpers";
 
 test("Should see the correct home page as a state user", async ({
   page,
